% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcds-score.R
\name{rcds_grade}
\alias{rcds_grade}
\title{Convert a 0-100 RCDS score to a letter grade}
\usage{
rcds_grade(score)
}
\arguments{
\item{score}{Numeric score in \verb{[0, 100]}.}
}
\value{
A single-character grade.
}
\description{
Convert a 0-100 RCDS score to a letter grade
}
