% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcds-components.R
\name{rcds_north_arrow}
\alias{rcds_north_arrow}
\title{RCDS north arrow (ggspatial wrapper)}
\usage{
rcds_north_arrow(location = "tr", canvas = "dark", style = NULL, ...)
}
\arguments{
\item{location}{Corner, e.g. \code{"tr"}.}

\item{canvas}{Canvas key.}

\item{style}{A ggspatial north-arrow style. Default \code{north_arrow_minimal}.}

\item{...}{Passed to \code{\link[ggspatial:annotation_north_arrow]{ggspatial::annotation_north_arrow()}}.}
}
\description{
Pre-styled \code{\link[ggspatial:annotation_north_arrow]{ggspatial::annotation_north_arrow()}}. Defaults to the minimal
orienteering style; restrained, never the dominant element.
}
