% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcds-palettes.R
\name{rcds_palettes}
\alias{rcds_palettes}
\title{Catalogue of available palettes}
\usage{
rcds_palettes()
}
\value{
A named list grouping palette names by family.
}
\description{
Catalogue of available palettes
}
