% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcds-palettes.R
\name{scale_fill_rcds_c}
\alias{scale_fill_rcds_c}
\alias{scale_color_rcds_c}
\title{RCDS continuous fill/colour scales}
\usage{
scale_fill_rcds_c(palette = "seq_blue", reverse = FALSE, ...)

scale_color_rcds_c(palette = "seq_blue", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Palette name (a sequential or diverging family).}

\item{reverse}{Reverse the ramp.}

\item{...}{Passed to \code{\link[ggplot2:scale_fill_gradientn]{ggplot2::scale_fill_gradientn()}} /
\code{\link[ggplot2:scale_color_gradientn]{ggplot2::scale_color_gradientn()}}.}
}
\description{
RCDS continuous fill/colour scales
}
