% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcds-palettes.R
\name{scale_fill_rcds_d}
\alias{scale_fill_rcds_d}
\alias{scale_color_rcds_d}
\title{RCDS discrete fill/colour scales}
\usage{
scale_fill_rcds_d(palette = "qual_brand", reverse = FALSE, ...)

scale_color_rcds_d(palette = "qual_brand", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Qualitative palette name.}

\item{reverse}{Reverse the order.}

\item{...}{Passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}} downstream.}
}
\description{
RCDS discrete fill/colour scales
}
